requirements.txt
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
python-dotenv>=0.19.0
datasets>=2.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0
scikit-learn>=1.3.0
joblib>=1.3.0 